refactr(walletProvider): wallet_validator, wallet_query_service로 로직 분리
